fix: make wails build the app
- the app couldn't run using `wails dev`. just fixed it. nothing more than moving and editing around files and existing code.
