+++ frontend/package.json.md5
@@ -1,1 +1,1 @@
-dbbcd64a9f1ff4ebf57456f5a60a7e02
+51f92c2ef520eceb54f7d124c9f2836c
