--- frontend/src/pages/notes/Lexical/plugins/TreeViewPlugin.tsx
@@ -1,19 +1,0 @@
-import React, { FC } from 'react';
-import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
-import { TreeView } from "@lexical/react/LexicalTreeView";
-
-const TreeViewPlugin: FC = () => {
-    const [editor] = useLexicalComposerContext();
-    return (
-        <TreeView
-            viewClassName="tree-view-output"
-            timeTravelPanelClassName="debug-timetravel-panel"
-            timeTravelButtonClassName="debug-timetravel-button"
-            timeTravelPanelSliderClassName="debug-timetravel-panel-slider"
-            timeTravelPanelButtonClassName="debug-timetravel-panel-button"
-            editor={editor}
-        />
-    );
-};
-
-export default TreeViewPlugin;
