+++ frontend/src/pages/notes/Lexical/utils/debounce.ts
@@ -1,4 +1,4 @@
-import {DebouncedFunctionType} from "../types";
+export type DebouncedFunctionType = (...params: any[]) => void;
 
 /** To help debounce the action of syncing editor's content changes to the external store. Usually, this would be the piece that updates your remote database
  * @see https://konstantinmuenster.medium.com/how-to-build-a-text-editor-with-lexical-and-react-27204c186d0f
